improvements: added an explicit type conversion

+++ app/src/main/cpp/RenderProgram.cpp
@@ -24,7 +24,7 @@ void RenderProgram::setupGraphic(GLint width, GLint height) {
     _transformTechnique = std::make_shared<Render::Techniques::Transform3dTechnique>();
     _transformTechnique->setModel(glm::vec3{0.0f});
     _transformTechnique->setView(glm::vec3(0.0f, 0.0f, -20.0f));
-    _transformTechnique->setProjection(45.0f, width, height, 0.1f, 100.0f);
+    _transformTechnique->setProjection(45.0f, (float)width, (float)height, 0.1f, 100.0f);
 
     _object->addTechnique(Render::Techniques::TRANSFORM, _transformTechnique);
     _object->enableTechnique(Render::Techniques::TRANSFORM);
